refactor: extract upload and deploy to lib modules
