+++ src/lib.rs
@@ -2,9 +2,11 @@
 extern crate text_io;
 
 pub mod commands;
+pub mod deploy;
 pub mod http;
 pub mod install;
 pub mod installer;
 pub mod settings;
 pub mod terminal;
+pub mod upload;
 pub mod util;
