+++ src/upload/form/mod.rs
@@ -19,6 +19,7 @@ use project_assets::ProjectAssets;
 use text_blob::TextBlob;
 use wasm_module::WasmModule;
 
+// these likely belong elsewhere, perhaps in a mod for project type specific behavior.
 use super::{krate, Package};
 
 pub fn build(
